rename file to `store` instead of `context` because it does more than that
