+++ src/lib/core/hooks.ts
@@ -2,7 +2,7 @@ import { useEventListener } from '@/hooks/useEventListener'
 import { useHover } from '@/hooks/useHover'
 import { isDefined, isDifferent, mapTuple } from '@/utils/data'
 import { CSSProperties, RefObject, useEffect, useRef } from 'react'
-import { setGlobalStyle, setRules, useStore, useStoreDispatch } from './context'
+import { setGlobalStyle, setRules, useStore, useStoreDispatch } from './store'
 import { Effect, Effects } from './effects'
 import { Shape, Shapes } from './shapes'
 import { computeStyle, defaultCursorStyles, GlobalStyle, Style } from './style'
