+++ src/lib/core/index.ts
@@ -1,6 +1,7 @@
 export { CursorProvider } from './components/Provider'
 export { Effects } from './effects'
 export type { Effect } from './effects'
+export { setGlobalStyle } from './store'
 export { useCursorStyle, useCursorStyleOnHover, useGlobalStyle, useHideSystemCursor } from './hooks'
 export { Shapes } from './shapes'
 export type { Style } from './style'
